Refactor task-queue-status to it's own component and use as a snippet
