+++ src/lib/pages/nexus-endpoint.svelte
@@ -1,13 +1,12 @@
 <script lang="ts">
-  import Alert from '$lib/holocene/alert.svelte';
+  import type { Snippet } from 'svelte';
+
   import Button from '$lib/holocene/button.svelte';
   import Copyable from '$lib/holocene/copyable/index.svelte';
   import Link from '$lib/holocene/link.svelte';
   import Markdown from '$lib/holocene/monaco/markdown.svelte';
   import { translate } from '$lib/i18n/translate';
-  import { getPollers } from '$lib/services/pollers-service';
   import type { NexusEndpoint as Endpoint } from '$lib/types/nexus';
-  import { pluralize } from '$lib/utilities/pluralize';
   import {
     routeForNamespace,
     routeForNexus,
@@ -18,29 +17,12 @@
   let {
     endpoint,
     editDisabled = false,
-  }: { endpoint: Endpoint; editDisabled?: boolean } = $props();
-
-  let pollerCount: number | undefined = $state(undefined);
-
-  const checkTaskQueue = async (endpoint: Endpoint) => {
-    const targetNamespace = endpoint?.spec?.target?.worker?.namespace;
-    const targetTaskQueue = endpoint?.spec?.target?.worker?.taskQueue;
-    if (targetNamespace && targetTaskQueue) {
-      try {
-        const { pollers } = await getPollers({
-          namespace: targetNamespace,
-          queue: targetTaskQueue,
-        });
-        pollerCount = pollers.length;
-      } catch (error) {
-        pollerCount = undefined;
-      }
-    }
-  };
-
-  $effect(() => {
-    checkTaskQueue(endpoint);
-  });
+    taskQueueStatus,
+  }: {
+    endpoint: Endpoint;
+    editDisabled?: boolean;
+    taskQueueStatus?: Snippet;
+  } = $props();
 </script>
 
 <div class="flex flex-col gap-8">
@@ -91,21 +73,7 @@
           </Link>
         </Copyable>
       </div>
-      {#if pollerCount !== undefined}
-        <Alert
-          intent={pollerCount > 0 ? 'success' : 'warning'}
-          title={pollerCount
-            ? 'Task Queue is Active'
-            : 'Task Queue is Inactive'}
-        >
-          <div class="flex w-full items-center justify-between">
-            <p>
-              {pollerCount}
-              {pluralize('Worker', pollerCount)}
-            </p>
-          </div></Alert
-        >
-      {/if}
+      {@render taskQueueStatus?.()}
     </div>
   </div>
   <div class="w-full xl:w-1/2">
